requirements.txt
discord.py==2.3.2
fastapi==0.110.0
loguru==0.7.2
opencv-contrib-python
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
roboflow==1.1.26
uvicorn==0.29.0
